feat: Add Kafka module and KafkaProducerService
This commit adds a new Kafka module and the KafkaProducerService class to the project. The Kafka module is responsible for providing the KafkaProducerService, which is used to connect to a Kafka broker and emit events to specified topics. This addition enables the application to integrate with Kafka and send messages to other services.

Ref: Issue #123